Basic scene rendering and movement working

+++ TheForceEngine/TFE_Jedi/Renderer/RClassic_Fixed/robj3d_fixed/robj3dFixed.cpp
@@ -22,7 +22,7 @@ namespace RClassic_Fixed
 {
 	void robj3d_projectVertices(vec3_fixed* pos, s32 count, vec3_fixed* out);
 	void robj3d_drawVertices(s32 vertexCount, const vec3_fixed* vertices, u8 color);
-	s32 polygonSort(const void* r0, const void* r1);
+	int polygonSort(const void* r0, const void* r1);
 
 	void robj3d_draw(SecObject* obj, JediModel* model)
 	{
@@ -113,11 +113,11 @@ namespace RClassic_Fixed
 		}
 	}
 
-	s32 polygonSort(const void* r0, const void* r1)
+	int polygonSort(const void* r0, const void* r1)
 	{
 		JmPolygon* p0 = *((JmPolygon**)r0);
 		JmPolygon* p1 = *((JmPolygon**)r1);
-		return p1->zAve - p0->zAve;
+		return (int)(p1->zAve - p0->zAve);
 	}
 
 }}  // TFE_Jedi
